refactor: split orchestrate page into multiple components
