+++ src/Router.tsx
@@ -3,19 +3,7 @@ import { useStore } from '@nanostores/react'
 import { GraphiQL } from 'graphiql'
 import { BrowserRouter, HashRouter, Route, Routes } from 'react-router-dom'
 
-import {
-  ConfigPage,
-  DNSPage,
-  ExperimentPage,
-  GroupPage,
-  HomePage,
-  MainLayout,
-  NodePage,
-  OrchestratePage,
-  RoutingPage,
-  SetupPage,
-  SubscriptionPage,
-} from '~/pages'
+import { ExperimentPage, HomePage, MainLayout, OrchestratePage, SetupPage } from '~/pages'
 import { endpointURLAtom } from '~/store'
 
 export const Router = () => {
@@ -27,12 +15,6 @@ export const Router = () => {
       <Routes>
         <Route path="/" element={<MainLayout />}>
           <Route index element={<HomePage />} />
-          <Route path="node" element={<NodePage />} />
-          <Route path="subscription" element={<SubscriptionPage />} />
-          <Route path="dns" element={<DNSPage />} />
-          <Route path="routing" element={<RoutingPage />} />
-          <Route path="group" element={<GroupPage />} />
-          <Route path="config" element={<ConfigPage />} />
           <Route path="orchestrate" element={<OrchestratePage />} />
           <Route path="experiment" element={<ExperimentPage />} />
         </Route>
